20240605 Long으로 id들 변수 변경 phoneNumber로 변수이름 변경 경로들 추가 엔티티당 컨트롤러랑 서비스 생성

+++ src/main/java/ypjs/project/domain/Category.java
@@ -13,7 +13,7 @@ public class Category {
 
     @Id @GeneratedValue
     @Column(name = "CATEGORY_ID")
-    private int CategoryId;
+    private Long CategoryId;
 
     @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
     private List<Item> items = new ArrayList<>();
